+++ src/animation/use-visual-element-animation.ts
@@ -1,34 +1,34 @@
 import {
-    ValueAnimationControls,
-    ValueAnimationConfig,
-} from "./ValueAnimationControls"
+    VisualElementAnimationControls,
+    AnimationControlsConfig,
+} from "./VisualElementAnimationControls"
 import { useContext, useEffect } from "react"
 import { MotionProps } from "../motion/types"
 import { MotionContext } from "../motion/context/MotionContext"
 import { useConstant } from "../utils/use-constant"
 import { PresenceContext } from "../components/AnimatePresence/PresenceContext"
+import { VisualElement } from "../render/VisualElement"
+import { checkShouldInheritVariant } from "../motion/utils/should-inherit-variant"
 
 /**
  * Creates an imperative set of controls to trigger animations.
  *
  * This allows a consolidated, uniform API for animations, to be triggered by other APIs like the `animate` prop, or the gesture handlers.
  *
- * @param values
- * @param props
- * @param ref
- * @param subscribeToParentControls
- *
  * @internal
  */
-export function useValueAnimationControls<P>(
-    config: ValueAnimationConfig,
+export function useVisualElementAnimation<P>(
+    visualElement: VisualElement,
     props: P & MotionProps,
-    subscribeToParentControls: boolean
+    config: AnimationControlsConfig
 ) {
+    const subscribeToParentControls = checkShouldInheritVariant(props)
     const { variants, transition } = props
     const parentControls = useContext(MotionContext).controls
     const presenceContext = useContext(PresenceContext)
-    const controls = useConstant(() => new ValueAnimationControls<P>(config))
+    const controls = useConstant(
+        () => new VisualElementAnimationControls<P>(visualElement, config)
+    )
 
     // Reset and resubscribe children every render to ensure stagger order is correct
     if (!presenceContext || presenceContext.isPresent) {