+++ src/behaviours/use-drag.ts
@@ -1,12 +1,11 @@
 import { useEffect, useContext } from "react"
-import { MotionValuesMap } from "../motion/utils/use-motion-values"
-import { ValueAnimationControls } from "../animation/ValueAnimationControls"
+import { VisualElementAnimationControls } from "../animation/VisualElementAnimationControls"
 import { MotionPluginContext } from "../motion/context/MotionPluginContext"
 import { DraggableProps } from "./types"
-import { ComponentDragControls } from "./ComponentDragControls"
+import { VisualElementDragControls } from "./VisualElementDragControls"
 import { useConstant } from "../utils/use-constant"
-import { NativeElement } from "../motion/utils/use-native-element"
-import { usePresence } from "../components/AnimatePresence/use-presence"
+import { useIsPresent } from "../components/AnimatePresence/use-presence"
+import { VisualElement } from "../render/VisualElement"
 
 /**
  * A hook that allows an element to be dragged.
@@ -20,15 +19,14 @@ import { usePresence } from "../components/AnimatePresence/use-presence"
  */
 export function useDrag(
     props: DraggableProps,
-    nativeElement: NativeElement<Element>,
-    values: MotionValuesMap,
-    controls: ValueAnimationControls
+    visualElement: VisualElement,
+    controls: VisualElementAnimationControls
 ) {
     const { dragControls: groupDragControls } = props
     const { transformPagePoint } = useContext(MotionPluginContext)
 
     const dragControls = useConstant(
-        () => new ComponentDragControls({ nativeElement, values, controls })
+        () => new VisualElementDragControls({ visualElement, controls })
     )
     dragControls.updateProps({ ...props, transformPagePoint })
 
@@ -39,15 +37,12 @@ export function useDrag(
         [dragControls]
     )
 
-    useEffect(() => dragControls.mount(nativeElement.getInstance()), [])
+    useEffect(() => dragControls.mount(visualElement.getInstance()), [])
 }
 
-function useDisableDragOnExit(dragControls: ComponentDragControls) {
-    const [isPresent, safeToRemove] = usePresence()
+function useDisableDragOnExit(dragControls: VisualElementDragControls) {
+    const isPresent = useIsPresent()
     useEffect(() => {
-        if (!isPresent && safeToRemove) {
-            safeToRemove()
-            dragControls.stopMotion()
-        }
+        if (!isPresent) dragControls.stopMotion()
     }, [isPresent])
 }