+++ src/motion/features/types.ts
@@ -1,18 +1,17 @@
 import * as React from "react"
 import { MotionProps } from "../types"
-import { ValueAnimationControls } from "../../animation/ValueAnimationControls"
-import { MotionValuesMap } from "../utils/use-motion-values"
+import { VisualElementAnimationControls } from "../../animation/VisualElementAnimationControls"
 import { MotionContextProps } from "../context/MotionContext"
-import { NativeElement } from "../utils/use-native-element"
 import { MotionPluginsContext } from "../context/MotionPluginContext"
+import { VisualElement } from "../../render/VisualElement"
+import { HTMLVisualElement } from "../../render/dom/HTMLVisualElement"
 
 /**
  * @public
  */
 export interface FeatureProps extends MotionProps {
-    controls: ValueAnimationControls
-    values: MotionValuesMap
-    nativeElement: NativeElement
+    visualElement: HTMLVisualElement
+    controls: VisualElementAnimationControls
     localContext: MotionContextProps
     parentContext: MotionContextProps
 }
@@ -30,20 +29,17 @@ export interface MotionFeature {
 }
 
 export type LoadMotionFeatures<P = {}> = (
-    nativeElement: NativeElement,
-    values: MotionValuesMap,
+    visualElenent: VisualElement,
     props: P & MotionProps,
     context: MotionContextProps,
     parentContext: MotionContextProps,
-    controls: ValueAnimationControls<P>,
+    controls: VisualElementAnimationControls<P>,
     inherit: boolean,
     plugins: MotionPluginsContext
 ) => React.ReactElement<FeatureProps>[]
 
 export type RenderComponent<P = {}> = (
-    nativeElement: NativeElement,
-    style: React.CSSProperties,
-    values: MotionValuesMap,
-    props: P,
-    isStatic?: boolean
-) => React.ReactElement
+    Component: string | React.ComponentType<P>,
+    props: MotionProps,
+    visualElement: VisualElement
+) => any