+++ src/render/dom/HTMLVisualElement.ts
@@ -1,0 +1,193 @@
+import { VisualElement } from "../VisualElement"
+import { BoundingBox2D, AxisBox2D } from "../../types/geometry"
+import {
+    convertBoundingBoxToAxisBox,
+    transformBoundingBox,
+} from "../../utils/geometry"
+import { ResolvedValues } from "../types"
+import { buildHTMLStyles } from "./utils/build-html-styles"
+import { DOMVisualElementConfig, TransformOrigin } from "./types"
+import { isTransformProp } from "./utils/transform"
+import { getDefaultValueType } from "./utils/value-types"
+
+/**
+ * A VisualElement for HTMLElements
+ */
+export class HTMLVisualElement<
+    E extends HTMLElement | SVGElement = HTMLElement
+> extends VisualElement<E> {
+    /**
+     * A mutable record of styles we want to apply directly to the rendered Element
+     * every frame. We use a mutable data structure to reduce GC during animations.
+     */
+    style: ResolvedValues = {}
+
+    /**
+     * A record of styles we only want to apply via React. This gets set in useMotionValues
+     * and applied in the render function. I'd prefer this to live somewhere else to decouple
+     * VisualElement from React but works for now.
+     */
+    reactStyle: ResolvedValues = {}
+
+    /**
+     * A mutable record of CSS variables we want to apply directly to the rendered Element
+     * every frame. We use a mutable data structure to reduce GC during animations.
+     */
+    vars: ResolvedValues = {}
+
+    /**
+     * A mutable record of transforms we want to apply directly to the rendered Element
+     * every frame. We use a mutable data structure to reduce GC during animations.
+     */
+    protected transform: ResolvedValues = {}
+
+    /**
+     * A mutable record of transform origins we want to apply directly to the rendered Element
+     * every frame. We use a mutable data structure to reduce GC during animations.
+     */
+    protected transformOrigin: TransformOrigin = {}
+
+    /**
+     * A mutable record of transform keys we want to apply to the rendered Element. We order
+     * this to order transforms in the desired order. We use a mutable data structure to reduce GC during animations.
+     */
+    protected transformKeys: string[] = []
+
+    protected defaultConfig: DOMVisualElementConfig = {
+        enableHardwareAcceleration: true,
+        allowTransformNone: true,
+    }
+
+    protected config = this.defaultConfig
+
+    /**
+     * The measured bounding box as it exists on the page with no transforms applied.
+     */
+    //private measuredBox = axisBox()
+
+    /**
+     * The `measuredBox` layout as corrected for all the transforms being applied up the
+     * tree. We use this as the final bounding box from which we calculate a transform
+     * delta to our desired visual position on any given frame.
+     *
+     * This is considered mutable to avoid object creation on each frame.
+     */
+    //private correctedBox = axisBox()
+
+    /**
+     * The visual target we want to project our component into on a given frame.
+     *
+     * This is considered mutable to avoid object creation on each frame.
+     */
+    //private targetBox = axisBox()
+
+    /**
+     * The overall scale of the local coordinate system as transformed by all parents of this component. We use this
+     * for scale correction on our calculated layouts and scale-affected values like `boxShadow`.
+     *
+     * This is considered mutable to avoid object creation on each frame.
+     */
+    //private treeScale = { x: 1, y: 1 }
+
+    /**
+     * When a value is removed, we want to make sure it's removed from all rendered data structures.
+     */
+    removeValue(key: string) {
+        super.removeValue(key)
+        delete this.vars[key]
+        delete this.style[key]
+    }
+
+    /**
+     * Empty the mutable data structures by re-creating them. We can do this every React render
+     * as the comparative workload to the rest of the render is very low and this is also when
+     * we want to reflect values that might have been removed by the render.
+     */
+    clean() {
+        this.style = {}
+        this.vars = {}
+        this.transform = {}
+    }
+
+    updateConfig(config: DOMVisualElementConfig = {}) {
+        this.config = { ...this.defaultConfig, ...config }
+    }
+
+    /**
+     * Measure and return the Element's bounding box. We convert it to a AxisBox2D
+     * structure to make it easier to work on each individual axis generically.
+     */
+    getBoundingBox(): AxisBox2D {
+        const { transformPagePoint } = this.config
+
+        let box = this.element.getBoundingClientRect() as BoundingBox2D
+        box = transformBoundingBox(box, transformPagePoint)
+        return convertBoundingBoxToAxisBox(box)
+    }
+
+    /**
+     * Build a style prop using the latest resolved MotionValues
+     */
+    build() {
+        // TODO: Add shadow bounding box resolution
+        buildHTMLStyles(
+            this.latest,
+            this.style,
+            this.vars,
+            this.transform,
+            this.transformOrigin,
+            this.transformKeys,
+            this.config
+        )
+    }
+
+    /**
+     * Read a value directly from the HTMLElement style.
+     */
+    read(key: string): number | string | null {
+        return this.getComputedStyle()[key] || 0
+    }
+
+    /**
+     * Read a value directly from the HTMLElement in case it's not defined by a Motion
+     * prop. If it's a transform, we just return a pre-defined default value as reading these
+     * out of a matrix is either error-prone or can incur a big payload for little benefit.
+     */
+    readNativeValue(key: string) {
+        if (isTransformProp(key)) {
+            const defaultValueType = getDefaultValueType(key)
+            return defaultValueType ? defaultValueType.default || 0 : 0
+        } else {
+            return this.read(key)
+        }
+    }
+
+    /**
+     * Return the computed style after a render.
+     */
+    getComputedStyle() {
+        return window.getComputedStyle(this.getInstance())
+    }
+
+    /**
+     * Render the Element by rebuilding and applying the latest styles and vars.
+     */
+    render() {
+        // Rebuild the latest animated values into style and vars caches.
+        this.build()
+
+        // Directly assign style into the Element's style prop. In tests Object.assign is the
+        // fastest way to assign styles.
+        Object.assign(this.element.style, this.style)
+
+        // Loop over any CSS variables and assign those.
+        for (const key in this.vars) {
+            this.element.style.setProperty(key, this.vars[key] as string)
+        }
+    }
+}
+
+// const axisBox = (): AxisBox2D => ({
+//     x: { min: 0, max: 0 },
+//     y: { min: 0, max: 0 },
+// })