+++ CHANGELOG.md
@@ -8,12 +8,14 @@ Framer Motion adheres to [Semantic Versioning](http://semver.org/).
 
 -   `AnimateSharedLayout`
 -   `onMeasureDragConstraints` prop that fires when `dragConstraints` are measured.
+-   `useIsPresent`, a read-only version of `usePresence`.
 
 ### Removed
 
 -   `positionTransition` and `layoutTransition` props (use `animate={true}` instead).
 -   `Point` type in favour of `Point2D`.
 -   `useAnimatedState`
+-   `stylefire` as a dependency.
 
 ### Fixed